Finished key listening
