+++ src/main/scala/it/unibo/pps1920/motoscala/view/screens/game/ScreenControllerGame.scala
@@ -1,11 +1,10 @@
-package it.unibo.pps1920.motoscala.view.screens
+package it.unibo.pps1920.motoscala.view.screens.game
 
 import it.unibo.pps1920.motoscala.controller.ObservableUI
 import it.unibo.pps1920.motoscala.controller.mediation.Event.{EntityData, LevelEndData, LevelSetupData}
 import it.unibo.pps1920.motoscala.controller.mediation.{Displayable, Event, Mediator}
 import it.unibo.pps1920.motoscala.view.ViewFacade
 import it.unibo.pps1920.motoscala.view.events.ViewEvent
-import it.unibo.pps1920.motoscala.view.screens.abstracts.AbstractScreenControllerGame
 
 class ScreenControllerGame(protected override val viewFacade: ViewFacade,
                            protected override val controller: ObservableUI)
